+++ CHANGES.md
@@ -4,7 +4,7 @@
     - `unikmer dump`: fix a nil pointer bug.
     - `unikmer count`:
         - fix checking taxid in sequence header.
-        - disable global taxid when using parsing taxid from sequence header.
+        - fix setting global taxid.
     - `unikmer count/diff/union`: slightly reduce memory and speedup when sorting k-mers.
 - v0.10.0
     - `unikmer`: fix loading custom taxonomy files.